longest drive done target zone done practice game some issue fix club change sometime issue fix.

+++ lib/feature/distance_control_drills/target_zone/data/model/shot_result.dart
@@ -1,6 +1,6 @@
 import 'package:equatable/equatable.dart';
 class ShotResult extends Equatable {
-  final int actualCarry;
+  final double actualCarry;
   final DateTime timestamp;
   final bool isWithinTarget;
